requirements.txt
easyocr
Pillow
pyperclip
keyboard
requests
pyqt5